Moves the RTP sequence number rewriter to the .rtp package.
