+++ src/org/jitsi/impl/neomedia/rtp/SequenceNumberRewriter.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jitsi.impl.neomedia;
+package org.jitsi.impl.neomedia.rtp;
 
+import org.jitsi.impl.neomedia.*;
 import org.jitsi.util.*;
 
 /**
